WIP Quest system port
This is a rough port of my quest interpreter I wrote for a personal project. This commit introduces a working system with a simple API to extend it. I haven't documented it fully yet, however. There might even be export inconsistencies across the code base now, as I haven't dealt with polishing yet.

New assets were also added, these are mostly improvements I made over time that are simply better to be backported to this repo.